requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.16
altair==5.3.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
blinker==1.8.2
blurhash==1.1.4
boost==0.1
bump2version==1.0.1
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
coremltools==7.2
coverage==7.5.4
cycler==0.12.1
dbscan==0.0.12
debugpy==1.8.2
decorator==5.1.1
deepface==0.0.92
distlib==0.3.8
distro==1.9.0
dlib==19.24.4
docutils==0.21.2
et-xmlfile==1.1.0
executing==2.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-videointelligence==2.13.4
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
gunicorn==22.0.0
h5py==3.11.0
httplib2==0.22.0
idna==3.7
imageio==2.34.2
imagesize==1.4.1
imutils==0.5.4
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
Mastodon.py==1.8.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.5
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
package_name==0.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==24.7.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.0
pycparser==2.22
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyproject-api==1.7.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
scikit-build==0.18.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==71.1.0
six==1.16.0
smmap==5.0.1
sounddevice==0.4.7
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
streamlit==1.37.0
streamlit-drawable-canvas==0.9.3
sympy==1.13.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-macos==2.16.2
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.7.21
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
